+++ HISTORY.rst
@@ -2,6 +2,39 @@
 History
 =======
 
+1.3.0 (2018-01-05)
+------------------
+
+* MV cut to account for the coverage (#119)
+* Remove MV cut from calibration data (#121)
+* Use correct position for S1PatternLikelihood cut (#123)
+* New S1 width cut to remove AC candidates (#124)
+* New cut on area fraction near Rn220 injection points (#125)
+* Tune S1 AFT cut based on bug fix and correct position (#126)
+* Revert to SR0 version of S2 tail cut (#128)
+* Optimise the PosDiff Cut (#127), interim update (#120)
+
+1.2.2 (2017-12-04)
+------------------
+
+* Update S1PatternLikelihood Definition (#101)
+* Add Flash cut to SR0 (#104)
+* Add flash cut variable to SR1 (#102)
+* Add cut for Misidentified krypton S1 (#103)
+* Add check that nearest flash isn't nan (#106)
+* Add parentheses (#107)
+* Try again with flasher cut (#108)
+* Remove S2 tail cut from calibration modes (#88)
+* Try again Flash cut (#109)
+* Update S2Width Cut (#96)
+* Change bounds on S1SingleScatter Cut (#113)
+* PosDiff Cut Definition (#111)
+* S2Width Warnings Fix (#114)
+* Add PosDiff; Remove S2Tails (duplicate); SR1 (#115)
+* Remove S2Tails (duplicate); SR0 (#116)
+* Fix S2Tails bug in cut lists (#117)
+
+
 1.2.1 (2017-11-27)
 ------------------